separate into multiple files
* independed type App -> app.go
* independed type APIAuthor -> app.go
* comment for the package -> kii.go
* remove "kii_cloud" prefix, because it is obvious for "package kii"
