+++ api.go
@@ -1,5 +1,3 @@
-// Package kii provides APIs to access to Kii Cloud and
-// Thing Interaction Framework (thing-if).
 package kii
 
 import (
@@ -10,42 +8,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"strings"
 )
-
-// App represents Application in Kii Cloud.
-type App struct {
-	AppID       string
-	AppKey      string
-	AppLocation string
-}
-
-// HostName returns host name of the Application endpoint.
-func (a *App) HostName() string {
-	lowerLoc := strings.ToLower(a.AppLocation)
-	switch lowerLoc {
-	case "jp":
-		return "api-jp.kii.com"
-	case "us":
-		return "api.kii.com"
-	case "cn":
-		return "api-cn3.kii.com"
-	case "sg":
-		return "api-sg.kii.com"
-	default:
-		return lowerLoc
-	}
-}
-
-// ThingIFBaseURL returns thing-if endpoint base url.
-func (a *App) ThingIFBaseURL() string {
-	return fmt.Sprintf("https://%s/thing-if/apps/%s", a.HostName(), a.AppID)
-}
-
-// KiiCloudBaseURL returns Kii Cloud endpoint base url.
-func (a *App) KiiCloudBaseURL() string {
-	return fmt.Sprintf("https://%s/api/apps/%s", a.HostName(), a.AppID)
-}
 
 // LayoutPosition represents Layout position of the Thing.
 type LayoutPosition int
@@ -120,13 +83,6 @@ type MqttEndpoint struct {
 	Password       string `json:"password"`
 	PortSSL        int    `json:"portSSL"`
 	PortTCP        int    `json:"portTCP"`
-}
-
-// Struct represents API author.
-// Can be Gateway, EndNode or KiiUser, depending on the token.
-type APIAuthor struct {
-	Token string
-	App   App
 }
 
 // Struct for requesting end node token
@@ -229,133 +185,4 @@ func AnonymousLogin(app App) (*APIAuthor, error) {
 		Token: respObj.AccessToken,
 		App:   app,
 	}, nil
-}
-
-// Let Gateway onboard to the cloud.
-// When there's no error, OnboardGatewayResponse is returned.
-func (au *APIAuthor) OnboardGateway(request OnboardGatewayRequest) (*OnboardGatewayResponse, error) {
-	reqJSON, err := json.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-	url := fmt.Sprintf("%s/onboardings", au.App.ThingIFBaseURL())
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(reqJSON))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/vnd.kii.onboardingWithVendorThingIDByThing+json")
-	req.Header.Set("Authorization", "Bearer "+au.Token)
-
-	bodyStr, err := executeRequest(req)
-	if err != nil {
-		return nil, err
-	}
-	var ret OnboardGatewayResponse
-	err = json.Unmarshal(bodyStr, &ret)
-	if err != nil {
-		return nil, err
-	}
-	return &ret, nil
-}
-
-// Request access token of end node of gateway.
-// Notes the APIAuthor should be a Gateway.
-// When there's no error, EndNodeTokenResponse is returned.
-func (au APIAuthor) GenerateEndNodeToken(gatewayID string, endnodeID string, request EndNodeTokenRequest) (*EndNodeTokenResponse, error) {
-	url := fmt.Sprintf("%s/things/%s/end-nodes/%s/token", au.App.KiiCloudBaseURL(), gatewayID, endnodeID)
-
-	reqJSON, err := json.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(reqJSON))
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+au.Token)
-
-	bodyStr, err := executeRequest(req)
-	if err != nil {
-		return nil, err
-	}
-	var ret EndNodeTokenResponse
-	err = json.Unmarshal(bodyStr, &ret)
-	if err != nil {
-		return nil, err
-	}
-	return &ret, nil
-}
-
-// Add an end node thing to gateway
-// Notes that the APIAuthor should be a Gateway
-func (au APIAuthor) AddEndNode(gatewayID string, endnodeID string) error {
-	url := fmt.Sprintf("%s/things/%s/end-nodes/%s", au.App.KiiCloudBaseURL(), gatewayID, endnodeID)
-
-	req, err := http.NewRequest("PUT", url, nil)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+au.Token)
-
-	if _, err := executeRequest(req); err != nil{
-		return err
-	}
-	return nil
-}
-
-// Register Thing.
-// The request must consist of the predefined fields(see RegisterThingRequest).
-// If you want to add the custom fileds, you can simply make RegisterThingRequest as anonymous field of your defined request struct, like:
-//  type MyRegisterThingRequest struct {
-//    RegisterThingRequest
-//    MyField1             string
-//  }
-// Where there is no error, RegisterThingResponse is returned
-func (au APIAuthor) RegisterThing(request interface{}) (*RegisterThingResponse, error) {
-	reqJSON, err := json.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-
-	url := fmt.Sprintf("%s/things", au.App.KiiCloudBaseURL())
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(reqJSON))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/vnd.kii.ThingRegistrationRequest+json")
-	req.Header.Set("X-Kii-AppID", au.App.AppID)
-	req.Header.Set("X-Kii-AppKey", au.App.AppKey)
-
-	bodyStr, err := executeRequest(req)
-	if err != nil {
-		return nil, err
-	}
-	var ret RegisterThingResponse
-	err = json.Unmarshal(bodyStr, &ret)
-	if err != nil {
-		return nil, err
-	}
-	return &ret, nil
-}
-
-// Update Thing state.
-// Notes that the APIAuthor should be already initialized as a Gateway or EndNode
-func (au APIAuthor) UpdateState(thingID string, request interface{}) error {
-
-	reqJSON, err := json.Marshal(request)
-	if err != nil {
-		return err
-	}
-
-	url := fmt.Sprintf("%s/targets/thing:%s/states", au.App.ThingIFBaseURL(), thingID)
-	req, err := http.NewRequest("PUT", url, bytes.NewBuffer(reqJSON))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+au.Token)
-
-	if _, err := executeRequest(req); err != nil {
-		return err
-	}
-	return nil
 }
