+++ kii.go
@@ -1,0 +1,3 @@
+// Package kii provides APIs to access to Kii Cloud and
+// Thing Interaction Framework (thing-if).
+package kii
